Moved scryfall query to data_fetch. Tweaked main.py so basic functionality works/
Scryfall query moved to data_fetch, and accessor function added to data_core for use in main. The function returns a dict with card info, and an error message.
Add some extra functionality to WUBRG to help with scryfall parsing.
Modified references and function in main to accomodate previous changes.